feat: compatible with server menus
