+++ arco-design-pro-vite/src/router/constans.ts
@@ -1,2 +1,10 @@
-// eslint-disable-next-line import/prefer-default-export
 export const DEFAULT_LAYOUT = () => import('@/layout/default-layout.vue');
+
+export const WHITE_LIST = [
+  { name: 'notFound', children: [] },
+  { name: 'login', children: [] },
+];
+
+export const NOT_FOUND = {
+  name: 'notFound',
+};
