+++ arco-design-pro-vite/src/router/guard/index.ts
@@ -1,7 +1,7 @@
 import type { Router } from 'vue-router';
 import { setRouteEmitter } from '@/utils/route-listener';
+import setupUserLoginInfoGuard from './userLoginInfo';
 import setupPermissionGuard from './permission';
-import setupServerPermissionGuard from './serverPermission';
 
 function setupPageGuard(router: Router) {
   router.beforeEach(async (to) => {
@@ -12,6 +12,6 @@ function setupPageGuard(router: Router) {
 
 export default function createRouteGuard(router: Router) {
   setupPageGuard(router);
-  setupServerPermissionGuard(router);
+  setupUserLoginInfoGuard(router);
   setupPermissionGuard(router);
 }
