+++ arco-design-pro-vite/src/store/modules/app/index.ts
@@ -68,6 +68,9 @@ const useAppStore = defineStore('app', {
         });
       }
     },
+    clearServerMenu() {
+      this.serverMenu = [];
+    },
   },
 });
