+++ arco-design-pro-vite/src/store/modules/app/types.ts
@@ -1,3 +1,5 @@
+import type { RouteRecordNormalized } from 'vue-router';
+
 export interface AppState {
   theme: string;
   colorWeak: boolean;
@@ -12,5 +14,6 @@ export interface AppState {
   device: string;
   tabBar: boolean;
   menuFromServer: boolean;
+  serverMenu: RouteRecordNormalized[];
   [key: string]: unknown;
 }
